requirements.txt
boto3
grpcio
grpcio-tools
awscli
hydra-core
#nlp workload
#tiktoken
#datasets
#wandb
#transformers
#export PYTHONPATH=/workspaces/super-dl